Update Action access control, outputs
